OrthoMPR subclass, instead of stringID
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@8348 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ ViewerController.h
@@ -608,7 +608,7 @@ enum
 - (void) loadSeries:(NSNumber*) t;
 - (void) offFullScreen;
 - (float) frame4DRate;
-- (long) maxMovieIndex;
+- (short) maxMovieIndex;
 - (NSSlider*) moviePosSlider;
 - (NSSlider*) sliderFusion;
 - (IBAction) convMatrixAction:(id)sender;
